[fix] 가격 전달 0원 버그 해결 (타입문제, null)
- 백엔드쪽 상품 응답 및 장바구니 아이템 DTO에서 가격 필드를 Double에서 BigDecimal로 변경

+++ commerce/backend/src/main/java/io/goorm/backend/dto/res/ProductResponse.java
@@ -3,6 +3,7 @@ package io.goorm.backend.dto.res;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -14,7 +15,7 @@ public class ProductResponse {
     private String link;
     private String image;
     private String author;
-    private Double discount;
+    private BigDecimal discount;
     private String publisher;
     private String pubdate;
     private String isbn;
